add `LatestConsumableOffsets` to `producer_worker`
And use it in the verifier workers.

In the case that the topic being consumed from had tombstones produced,
the high watermark may be given by a tombstone record that has been removed.
In trying to consume until this point, sequential readers will become stuck
polling for new records. Persist the last consumable offset in order to
adjust the offset we attempt to read up to in the read workers.